FastApi_3/Task 6.4/requirements.txt
fastapi
uvicorn
python-dotenv
pyjwt
passlib[bcrypt]